修正　Matching：完成説濃厚

+++ resources/views/games/matching.blade.php
@@ -33,6 +33,7 @@
             .then(data => {
                 // 部屋が満員かどうかを確認
                 if (data.isFull) {
+                    isAutoRedirect = true;
                     // 部屋が満員になったらプレイ画面にリダイレクト
                     window.location.href = '/gameroom/{{ $room->id }}';
                 } else {
@@ -45,7 +46,7 @@
     }, 500); // 1秒ごとにサーバーの状態を確認
 
     window.addEventListener('beforeunload', (event) => {
-        if (!isAutoRedirect || window.location.href != '/gameroom/{{ $room->id }}') {
+        if (!isAutoRedirect && window.location.href !== '/gameroom/{{ $room->id }}') {
             fetch(`{{route('removeMatchingRoom')}}`, {
                 method: 'DELETE',
                 headers: {
@@ -61,6 +62,10 @@
             });
         }
 
+    });
+
+    window.addEventListener('load', () => {
+        isAutoRedirect = false; // ページが読み込まれたら元に戻す
     });
 </script>
